[-] alter timetable.execution_log.txid column type to bigint, fixes #560

+++ internal/pgengine/pgengine_test.go
@@ -142,7 +142,7 @@ func TestSchedulerFunctions(t *testing.T) {
 		var chains []pgengine.ChainTask
 		tx, txid, err := pge.StartTransaction(ctx, 0)
 		assert.NoError(t, err, "Should start transaction")
-		assert.Greater(t, txid, 0, "Should return transaction id")
+		assert.Greater(t, txid, int64(0), "Should return transaction id")
 		assert.NoError(t, pge.GetChainElements(ctx, &chains, 0), "Should no error in clean database")
 		assert.Empty(t, chains, "Should be empty in clean database")
 		pge.CommitTransaction(ctx, tx)
@@ -152,7 +152,7 @@ func TestSchedulerFunctions(t *testing.T) {
 		var paramVals []string
 		tx, txid, err := pge.StartTransaction(ctx, 0)
 		assert.NoError(t, err, "Should start transaction")
-		assert.Greater(t, txid, 0, "Should return transaction id")
+		assert.Greater(t, txid, int64(0), "Should return transaction id")
 		assert.NoError(t, pge.GetChainParamValues(ctx, &paramVals, &pgengine.ChainTask{
 			TaskID:  0,
 			ChainID: 0}), "Should no error in clean database")
@@ -170,7 +170,7 @@ func TestSchedulerFunctions(t *testing.T) {
 	t.Run("Check ExecuteSQLCommand function", func(t *testing.T) {
 		tx, txid, err := pge.StartTransaction(ctx, 0)
 		assert.NoError(t, err, "Should start transaction")
-		assert.Greater(t, txid, 0, "Should return transaction id")
+		assert.Greater(t, txid, int64(0) , "Should return transaction id")
 		f := func(sql string, params []string) error {
 			_, err := pge.ExecuteSQLCommand(ctx, tx, sql, params)
 			return err
